Handle CSTACK_ADDRESS of MOST_NEGATIVE_FIX

+++ gcl/h/notcomp.h
@@ -253,7 +253,7 @@ EXTER struct printStruct *printStructBufp;
 #ifndef NULL_OR_ON_C_STACK
 
 
-#if ((long)CSTACK_ADDRESS > 0)
+#if (CSTACK_ADDRESS > 0)
 #define NULL_OR_ON_C_STACK(x) (((unsigned long)x)<=DBEGIN || ((unsigned long)x) >= (unsigned long)(pagetochar(MAXPAGE)))
 #else
 #define NULL_OR_ON_C_STACK(x) ((long)x <= DBEGIN)
